+++ RELEASE_NOTES.md
@@ -1,3 +1,10 @@
+### Version 1.2.5
+__Changes__
+- removed manual configuration of phred33/phred64 quality encoding to make auto-detect
+- removed manual configuration of PE/SE to make auto-detect
+- added option for translating output to phred33
+- changed adapter option from TruSeq2-PE-2 to TruSeq3-PE-2
+
 ### Version 1.2.4
 __Changes__
 - removed bad formatting in output HTML